==> original_prereg/01.get_person_ids.Rmd <==
---
title: "Get Person Ids"
author: "Erin M. Buchanan"
date: "Last Knitted: `r Sys.Date()`"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Libraries

```{r}
library(dplyr)
library(purrr)
library(bib2df)
library(tidyverse)
library(httr)
library(googlesheets4)

library(rorcid)
# orcid_auth() run the first time to set up
# copy the bearer number 
# file.edit("~/.Renviron")
# ORCID_TOKEN="numberhere"

library(scholar)
library(rio)
```

## Import the BibTeX File

In this section, we are importing the bibtex file that includes all the papers we are going to use. 

Notes:

  - We used the preprint manuscript for some papers because they had the entire author list, we will update their references manually for the final dataset. 
  - We will have to manually add the names for several papers that used consortium authors. 
  - The pre-reg of this document only examines a few authors/papers to ensure code can be run and develop workflow. 

```{r eval = F}
# bibs <- bib2df("papers.bib")
bibs <- bib2df("manuscript_data/papers_06_17_22.bib")
```

## ORCID

In this section, we will figure out the ORCIDs for each author if possible. 

  - Data is stored in `bibs$AUTHOR` as a list. 
  - Extract that information and loop over the author list. 
  - Because authors will repeat, we need to put this in long format with one row per author and all other information saved. 
  - Next, split author name, so we can look it up with orcid.

```{r eval = F}
# get data into long format
long_bib <- bibs %>% unnest_longer(col = AUTHOR)

# remove this nonsense
long_bib$AUTHOR <- gsub("[{}]", "", long_bib$AUTHOR)

# now split the names into last, first middle 
# first split on the comma only 
long_bib <- long_bib %>% separate(col = AUTHOR, sep = ",", 
                                  into = c("last", "first"), 
                                  extra = "merge")

# remove the first space
long_bib$first <- gsub("^ ", "", long_bib$first)

# then split on space
long_bib <- long_bib %>% separate(col = first, sep = " ", 
                                  into = c("first", "middle"),
                                  extra = "merge")

# drop editor column so it is a non-nested DF
long_bib <- long_bib %>% select(-EDITOR)
write.csv(long_bib, "id_output/unedited_bib.csv", row.names = F)
```

To make the finalized data:
  
  - Loop over rows and find candidates for ORCIDs.
  - Export the candidate data.
  - Export the bibtex data (done above).
  - A bit of fun manual merging. 

```{r eval = F}
# there are going to be a lot of duplicates
# really only need to check them once if exact match
# not perfect exclusion because of symbol differences and middles but ok 
get_id <- long_bib %>% select(last, first, middle, TITLE)
get_id <- get_id %>% filter(!duplicated(get_id[ , c("last", "first")]))

# clean up titles
get_id$TITLE <- gsub("[[:punct:]]", "", get_id$TITLE)
get_id$TITLE <- tolower(get_id$TITLE)

get_id$ORCID <- NA

for (i in 1:nrow(get_id)){
#for (i in 1:10){ #temporary do a small number 
  
  # find possible candidates 
  possibles <- orcid_search(family_name = get_id$last[i], 
                            given_name = get_id$first[i])
  
  cat(i)
  cat("\n")
  
  # if nrow = 0 then move on 
  
  # if nrow == 1 then figure out if there's a match
  if (nrow(possibles) == 1){
    temp_works <- orcid_works(possibles$orcid) 
    temp_DF <- do.call(bind_rows, temp_works)
    
    if (nrow(temp_DF) > 0){
    temp_DF$title.title.value <- gsub("[[:punct:]]", "", temp_DF$title.title.value) 
    temp_DF$title.title.value <- tolower(temp_DF$title.title.value) 
    temp_DF <- temp_DF %>% 
      filter(temp_DF$title.title.value == get_id$TITLE[i])
    
    id <- unique(na.omit(temp_DF$`source.assertion-origin-orcid.path`))
    
    if(length(id) == 1) { get_id$ORCID[i] <- id }
    if(length(id) > 1) { get_id$ORCID[i] <- list(id) }
    
    }
  }
  
  # if nrow > 1 figure it out which matches
  if (nrow(possibles) > 1){
    temp_works <- orcid_works(possibles$orcid) 
    temp_DF <- do.call(bind_rows, temp_works)
    
    if (nrow(temp_DF) > 0){
    temp_DF$works$title.title.value <- gsub("[[:punct:]]", "", temp_DF$works$title.title.value) 
    temp_DF$works$title.title.value <- tolower(temp_DF$works$title.title.value) 
    temp_DF <- temp_DF %>% 
      filter(temp_DF$works$title.title.value == get_id$TITLE[i])
    
    id <- unique(na.omit(temp_DF$works$`source.assertion-origin-orcid.path`))
    
    if(length(id) == 1) { get_id$ORCID[i] <- id }
    if(length(id) > 1) { get_id$ORCID[i] <- list(id) }
    
    }
  }
  
  # be nice to api take a break 
  Sys.sleep(runif(1, 3, 5))
  
}

write.csv(get_id, "id_output/possible_orc_ids.csv", row.names = F)
```

## Google Scholar

- First, fix the scholar package issue that it doesn't correctly pull ids that have punctuation in them.

```{r}
get_scholar <- function (last_name = "", first_name = "", affiliation = NA) 
{
    if (!any(nzchar(c(first_name, last_name)))) 
        stop("At least one of first and last name must be specified!")
    site <- getOption("scholar_site")
    url <- paste0(site, "/citations?view_op=search_authors&mauthors=", 
        first_name, "+", last_name, "&hl=en&oi=ao")
    page <- get_scholar_resp(url)
    if (is.null(page)) 
        return(NA)
    aa <- content(page, as = "text")
    ids <- stringr::str_extract_all(string = aa, pattern = ';user=.{12}')
    
    if (length(unlist(ids)) == 0) {
        message("No Scholar ID found.")
        return(NA)
    }
    ids <- ids %>% unlist %>% gsub(";user=", "", 
        .) %>% unique
    
    return(ids)
  }
```

- Then determine if you can find scholars matching that information. 

```{r eval = F}
# thanks SO for this id for the bad coding on scholar 
p_get_scholar_id <- possibly(get_scholar, otherwise = NA_character_)
scholars <- character(nrow(get_id))

for(i in 1:length(scholars)) {
  
  id <- NA
  possibles <- p_get_scholar_id(last_name = get_id$last[i], 
                                first_name = get_id$first[i])
  
  for (q in 1:length(possibles)){
    
    articles <- get_publications(possibles[q])
    articles$title <- gsub("[[:punct:]]", "", articles$title) 
    articles$title <- tolower(articles$title) 
    temp_DF <- articles %>% 
      filter(title == get_id$TITLE[i])
    
    if (nrow(temp_DF) > 0){
      
      if (!is.na(id)){ 
        id <- c(id, possibles[q]) 
        } else { id <- possibles[q] }
    }
    
  }
  cat(i)
  cat("\n")
  
  if (length(id) == 1) { scholars[i] <- id }
  if (length(id) > 1) { scholars[i] <- list(id) }
  
  # be nice to api take a break 
  Sys.sleep(runif(1, 3, 5))
}

get_id$scholar <- scholars

write.csv(get_id, "id_output/possible_both_ids.csv", row.names = F)
```

## Get Data Round 2

```{r}
DF <- read_sheet("https://docs.google.com/spreadsheets/d/1dSlIQXeLcFgkITyfy15_yAdtjct1AedjD6s5D4qTzCw/edit#gid=1698777071")
```

## Work on ORC 

```{r eval = F}
ORC <- DF %>% 
  filter(ORCID == "FIND")

for (i in 1:nrow(ORC)){
#for (i in 1:10){ #temporary do a small number 
  
  # find possible candidates 
  possibles <- orcid_search(family_name = ORC$last[i], 
                            given_name = ORC$first[i])
  
  cat(i)
  cat("\n")
  
  # if nrow == 1 then figure out if there's a match
  if (nrow(possibles) == 1){
    ORC$ORCID[i] <- possibles$orcid
  }
  
  # be nice to api take a break 
  Sys.sleep(runif(1, 3, 5))
  
}

DF <- DF %>% 
  full_join(ORCF %>% select(first, last, ORCID), 
            by = c("first" = "first", "last" = "last"))

export(DF, "id_output/orc_final.xlsx")
```

## Get Scholar IDs

```{r}
scholars <- DF %>% filter(scholar == "FIND")
p_get_scholar_id <- possibly(get_scholar, otherwise = NA_character_)

nrow(scholars)
# start at Bergmann 134

for(i in 3601:nrow(scholars)) {
  
  possibles <- p_get_scholar_id(last_name = scholars$last[i], 
                                first_name = scholars$first[i])
  
  if (length(possibles) == 1) { scholars$scholar[i] <- possibles }
  
  # be nice to api take a break 
  Sys.sleep(runif(5, 3, 5))
  
  # if divisible by 10
  if (i %% 10 == 0){
    export(scholars, "id_output/scholar_update.xlsx")
  }
}

DF <- DF %>% 
  full_join(scholars %>% select(first, last, scholar), 
            by = c("first" = "first", "last" = "last"))

export(DF, "id_output/scholar_final.xlsx")

```
