==> original_prereg/02.get_person_stats.Rmd <==
---
title: "Get Person Ids"
author: "Erin M. Buchanan"
date: "Last Knitted: `r Sys.Date()`"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Libraries

```{r}
library(dplyr)
library(rio)
library(tidyr)
library(rorcid)
library(purrr)
library(anytime)
library(janitor)
```

## Import Unique People

In trying to run our `01.get_person_ids.Rmd` file, we figured out that a lot of the data did not bring up matches. Therefore, we coded most scholar and orcids by hand.

```{r}
id_list <- import("test_utf8.xlsx")
```

## Get Person Statistics 

### ORCID

```{r}
id_list$ORCID <- gsub("^NA$", NA, id_list$ORCID)

orc_list <- na.omit(unique(id_list$ORCID))
author_works <- list()
author_ed <- list()
author_job <- list()

for (ORCID in orc_list) {
  
  # get works
  works_df <- works(ORCID) %>%
  as_tibble() %>%
  janitor::clean_names() %>%
  dplyr::mutate(created_date_value = anytime::anydate(created_date_value/1000))

  # make all works normalized
  works_df$title_title_value <- tolower(works_df$title_title_value)
  works_df$title_title_value <- gsub("[[:punct:]]", " ", works_df$title_title_value)

  # remove real duplicates
  author_works[[ORCID]] <- works_df[!duplicated(works_df[ , c("our_type", "title_title_value") ]), ]
  
  Sys.sleep(runif(1, 1, 2))
  
  # education 
  author_ed[[ORCID]] <- orcid_educations(ORCID) %>% 
    purrr::map(., purrr::pluck, "affiliation-group", "summaries") %>% 
    purrr::flatten_dfr() %>% 
    janitor::clean_names() 
    
  Sys.sleep(runif(1, 1, 2))
  
  # employments
  author_job[[ORCID]] <- orcid_employments(ORCID) %>%
  purrr::map(., purrr::pluck, "affiliation-group", "summaries") %>% 
  purrr::flatten_dfr() %>%
  janitor::clean_names() 
  Sys.sleep(runif(1, 1, 2))
  
}

works_df <- do.call(bind_rows, author_works) %>% select(-external_ids_external_id)
works_df$ORCID <- rep(names(author_works), lapply(author_works, nrow))
works_df <- works_df %>% 
      dplyr::mutate(last_modified_date_value = 
                      anytime::anydate(last_modified_date_value/1000))

  
write.csv(works_df, "works_output/ORCID_works.csv", row.names = F)

ed_df <- do.call(bind_rows, author_ed)
ed_df$ORCID <- rep(names(author_ed), lapply(author_ed, nrow))
ed_df <- ed_df %>% 
    dplyr::mutate(education_summary_created_date_value =
                  anytime::anydate(education_summary_created_date_value/1000),
                education_summary_last_modified_date_value =
                  anytime::anydate(education_summary_last_modified_date_value/1000))

write.csv(ed_df, "works_output/ORCID_ed.csv", row.names = F)

job_df <- do.call(bind_rows, author_job)
job_df$ORCID <- rep(names(author_job), lapply(author_job, nrow))
job_df <- job_df %>%
  dplyr::mutate(employment_summary_end_date =
                  anytime::anydate(employment_summary_end_date/1000),
                employment_summary_created_date_value =
                  anytime::anydate(employment_summary_created_date_value/1000),
                employment_summary_last_modified_date_value =
                  anytime::anydate(employment_summary_last_modified_date_value/1000))
write.csv(job_df, "works_output/ORCID_job.csv", row.names = F)
```

### Scholar 

```{r}
id_list$scholar <- gsub("^NA$", NA, id_list$scholar)

scholar_list <- na.omit(unique(id_list$scholar))
s_author_works <- list()
s_author_info <- list()

for (scholar in scholar_list) {
  
  # works 
  works_df <- get_publications(scholar)
  works_df$title <- tolower(works_df$title)
  works_df$title <- gsub("[[:punct:]]", " ", works_df$title)
  
  Sys.sleep(runif(1, 1, 2))
  
  # remove duplicates
  s_author_works[[scholar]] <- works_df %>% 
    arrange(desc(cites)) %>% 
    filter(!duplicated(works_df %>% select(title)))
  
  # information
  s_author_info[[scholar]] <- get_profile(scholar)
  
  Sys.sleep(runif(1, 1, 2))
  
}

works_df <- do.call(bind_rows, s_author_works) 
works_df$scholar <- rep(names(s_author_works), lapply(s_author_works, nrow))
write.csv(works_df, "works_output/scholar_works.csv", row.names = F)

# paste together information 
for (scholar in scholar_list){ 
  
  s_author_info[[scholar]]$interests <- paste(s_author_info[[scholar]]$interests, collapse = ";;")
  s_author_info[[scholar]]$coauthors <- paste(s_author_info[[scholar]]$coauthors, collapse = ";;")
  
  if (length(s_author_info[[scholar]]$homepage) < 1){
    s_author_info[[scholar]]$homepage <- NA
  }
  
  }

info_df <- do.call(bind_rows, s_author_info) 
info_df$scholar <- info_df$id
write.csv(info_df, "works_output/scholar_info.csv", row.names = F)
```
